Moved loading of grid variables into LoadData script.
Split transport models out from Transect and Grid Modeling script.

Now there is 1) LoadData file which brings in both transect and grid files and cleans them up. 2) Transect and grid modeling (which I want to change the name of) which is modeling the transect data. and 3) Transport models, which takes the transect models and applies to wind, water transport etc.